JepetoWoodwork/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
debugpy==1.6.7
Django==4.2.3
django-phonenumber-field==7.1.0
greenlet==2.0.2
gunicorn==21.2.0
idna==3.4
msgpack==1.0.5
packaging==23.1
phonenumberslite==8.13.15
Pillow==10.0.0
psycopg2-binary==2.9.6
pynvim==0.4.3
python-dotenv==1.0.0
requests==2.31.0
sqlparse==0.4.4
Unidecode==1.3.6
urllib3==2.0.3
